refactor(gcf-utils): separate integration tests into GCFLogger and GCFBootstrapper test files
